requirements/production-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --extra=production --output-file=requirements/production-requirements.txt --strip-extras
#
asgiref==3.9.1
    # via django
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
crispy-bootstrap5==2025.6
    # via kampanile3 (pyproject.toml)
directorium==0.2.0
    # via kampanile3 (pyproject.toml)
django==5.2.5
    # via
    #   crispy-bootstrap5
    #   django-crispy-forms
    #   kampanile3 (pyproject.toml)
django-crispy-forms==2.4
    # via
    #   crispy-bootstrap5
    #   kampanile3 (pyproject.toml)
django-npm==1.0.1
    # via kampanile3 (pyproject.toml)
django-split-settings==1.3.2
    # via kampanile3 (pyproject.toml)
idna==3.10
    # via requests
mido==1.3.3
    # via kampanile3 (pyproject.toml)
packaging==25.0
    # via mido
paho-mqtt==2.1.0
    # via kampanile3 (pyproject.toml)
platformdirs==4.3.8
    # via directorium
python-rtmidi==1.5.8
    # via kampanile3 (pyproject.toml)
requests==2.32.4
    # via directorium
sqlparse==0.5.3
    # via django
typing-extensions==4.14.1
    # via directorium
tzdata==2025.2
    # via django
urllib3==2.5.0
    # via requests
